GNSSユーティリティクラスの追加
